refactor(frontend): rename files
